redteamtool/requirements.txt
# Core Dependencies
click>=8.1.3
colorama>=0.4.6
rich>=13.6.0
pyyaml>=6.0.1

# Network Tools
scapy>=2.5.0
impacket>=0.11.0
python-nmap>=0.7.1
paramiko>=3.3.1

# Active Directory Tools
ldap3>=2.9.1
bloodhound>=1.4.0

# Web Tools
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.11.2

# Reporting
jinja2>=3.1.2
reportlab>=4.0.7
markdown>=3.4.3
weasyprint>=60.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Security
cryptography>=41.0.7
pyOpenSSL>=23.3.0

# Additional dependencies
dnspython>=2.4.2
python-docx>=1.0.1
